Struct ears::Music

pub struct Music { /* fields omitted */ }

Play Music easily.

Simple class to play music easily in 2 lines.

Music is played in their own task and the samples are loaded progressively using circular buffers.

Music maintains it’s own cursor into the underlying file and will use that cursor to continuously refill each buffer as it’s processed by the source.

They aren’t associated to a SoundData like Sounds.

Examples

extern crate ears;
use ears::{Music, SoundError, AudioController};

fn main() -> Result<(), SoundError> {
  let mut msc = Music::new("path/to/music.flac")?;
  msc.play();
  Ok(())
}

Implementations

impl Music

pub fn new(path: &str) -> Result<Music, SoundError>

Create a new Music

Argument

path - The path of the file to load the music

Return

A Result containing Ok(Music) on success, Err(SoundError) if there has been an error.