Struct ears::Sound

pub struct Sound { /* fields omitted */ }

Play Sounds easily.

Simple class to play sounds easily in 2 lines. Sounds are really light objects, the sound’s data is entirely loaded into memory and can be shared between Sounds using the SoundData object.

Examples

extern crate ears;
use ears::{Sound, AudioController, SoundError};

fn main() -> Result<(), SoundError> {
   // Create a Sound with the path of the sound file.
   let mut snd = Sound::new("path/to/my/sound.ogg")?;

   // Play it
   snd.play();

   // Wait until the sound stopped playing
   while snd.is_playing() {}

   Ok(())
}

Implementations

impl Sound

pub fn new(path: &str) -> Result<Sound, SoundError>

Default constructor for Sound struct.

Create a new struct and an associated SoundData.

Argument

path - The path of the sound file to create the SoundData.

Return

A Result containing Ok(Sound) on success, Err(SoundError) if there has been an error.

Example

use ears::Sound;

let snd = Sound::new("path/to/the/sound.ogg")
                 .expect("Cannot load the sound from a file!");

pub fn new_with_data(
sound_data: Arc<Mutex<SoundData>>
) -> Result<Sound, SoundError>

Create a new struct with a SoundData to associate.

Argument

sound_data - The sound_data to associate to the Sound.

Return

An Option with Some(Sound) if the Sound is created properly, or None if un error has occured.

Example

use ears::{Sound, SoundData, SoundError, AudioController};
use std::sync::{Mutex, Arc};

fn main() -> Result<(), SoundError> {
    let data = Arc::new(Mutex::new(SoundData::new("path/to/the/sound.ogg")?));
    let sound = Sound::new_with_data(data)?;
    Ok(())
}

pub fn get_datas(&self) -> Arc<Mutex<SoundData>>

Get the sound datas.

Return

The SoundData associated to this Sound.

Example

fn main() -> Result<(), ears::SoundError> {
    let snd = ears::Sound::new("path/to/the/sound.ogg")?;
    let snd_data = snd.get_datas();
    Ok(())
}

pub fn set_datas(&mut self, sound_data: Arc<Mutex<SoundData>>)

Set the sound datas.

Doesn’t work if the sound is currently playing.

Argument

sound_data - The new sound_data

Example

fn main() -> Result<(), ears::SoundError> {
    let snd1 = ears::Sound::new("path/to/the/sound.ogg")?;
    let mut snd2 = ears::Sound::new("other/path/to/the/sound.ogg")?;
    let snd_data = snd1.get_datas();
    snd2.set_datas(snd_data);
    Ok(())
}

pub fn set_air_absorption_factor(&mut self, factor: f32)

This is a multiplier on the amount of Air Absorption applied to the Source.

The air absorption factor is multiplied by an internal Air Absorption Gain

HF value of 0.994 (-0.05dB) per meter which represents normal atmospheric

humidity and temperature.

By default the value is set to 0.0 which means that Air Absorption effects

are disabled.

A value of 1.0 will tell the Effects Extension engine to apply high frequency

attenuation on the direct path of the Source at a rate of 0.05dB per meter.

Range 0.0 to 10.0

pub fn get_air_absorption_factor(&mut self) -> f32

Returns the current air absorption factor for the source.

pub fn set_velocity(&mut self, velocity: [f32; 3])

Set the velocity of a Sound.

Default velocity is [0.0, 0.0, 0.0].

Argument

velocity - A three dimensional vector of f32 containing the velocity of the sound [x, y, z].

pub fn get_velocity(&self) -> [f32; 3]

Get the velocity of a Sound.

Return

A three dimensional vector of f32 containing the velocity of the sound [x, y, z].