Struct ears::SoundData

pub struct SoundData { /* fields omitted */ }

Samples extracted from a file.

SoundDatas are made to be shared between several Sound and played in the same time.

Example

use ears::{Sound, SoundData, SoundError, AudioController};
use std::cell::RefCell;
use std::rc::Rc;

fn main() -> Result<(), SoundError> {
  // Create a SoundData
  let snd_data = Rc::new(RefCell::new(SoundData::new("path/to/my/sound.wav")?));

  // Create two Sound with the same SoundData
  let mut snd1 = Sound::new_with_data(snd_data.clone())?;
  let mut snd2 = Sound::new_with_data(snd_data.clone())?;

  // Play the sounds
  snd1.play();
  snd2.play();

  // Wait until snd2 is playing
  while snd2.is_playing() {}
  Ok(())
}

Implementations

impl SoundData

pub fn new(path: &str) -> Result<SoundData, SoundError>

Create a new SoundData.

The SoundData contains all the information extracted from the file: samples and tags. It’s an easy way to share the same samples between man Sounds objects.

Arguments

path - The path of the file to load

Return

A Result containing Ok(SoundData) on success, Err(SoundError) if there has been an error.