Struct ears::Music

pub struct Music {
    // some fields omitted
}

Play Music easily.

Simple class to play music easily in 2 lines.

Music is played in their own task and the samples are loaded progressively using circular buffers. They aren't associated to a SoundData like Sounds.

Examples

extern crate ears;
use ears::{Music, AudioController};

fn main() -> () {
  let mut msc = Music::new("path/to/music.flac").unwrap();
  msc.play();
}

Methods

impl Music

fn new(path: &str) -> Option<Music>

Create a new Music

Argument

path - The path of the file to load the music

Return

An Option containing Some(Music) on success, None otherwise